+++ lomond/stream.py
@@ -12,8 +12,9 @@ import logging
 from six import text_type
 
 from . import errors
-from .frame_parser import ParseError, FrameParser
+from .frame_parser import FrameParser
 from .message import Message
+from .parser import ParseError
 from .response import Response